requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via starlette
arabic-reshaper==3.0.0
    # via xhtml2pdf
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   gtts
    #   uvicorn
colorama==0.4.6
    # via click
fastapi==0.95.0
    # via -r requirements.in
gtts==2.3.2
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
html5lib==1.1
    # via xhtml2pdf
idna==3.4
    # via
    #   anyio
    #   requests
pillow==7.0.0
    # via
    #   -r requirements.in
    #   reportlab
    #   xhtml2pdf
pydantic==1.10.7
    # via fastapi
pypdf2==3.0.1
    # via xhtml2pdf
python-bidi==0.4.2
    # via xhtml2pdf
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
reportlab==3.6.6
    # via
    #   -r requirements.in
    #   xhtml2pdf
requests==2.28.2
    # via
    #   -r requirements.in
    #   gtts
six==1.16.0
    # via
    #   html5lib
    #   python-bidi
    #   xhtml2pdf
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   pydantic
    #   pypdf2
    #   starlette
urllib3==1.26.15
    # via requests
uvicorn==0.21.1
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
xhtml2pdf==0.2.5
    # via -r requirements.in
